feat: consolidated tip setting calls into a single call
