+++ packages/coil-client/src/queries/index.ts
@@ -6,4 +6,4 @@ export * from './adaptedPage'
 export * from './featureEnabled'
 export * from './tipPreview'
 export * from './tip'
-export * from './minTipLimit'
+export * from './tipSettings'
