+++ packages/coil-client/src/queries/tip.ts
@@ -8,31 +8,11 @@ export const tipQuery = `mutation tip($input: TipInput!) {
   }
 }`
 
-export const tipSettingsQuery = `
-  query updateTipSettings {
-    whoami {
-      tipping {
-        lastTippedAmount
-         limitRemaining
-      }
-    }
-  }
-`
-
 export interface TipData {
   tip: {
     code: string
     message: string
     success: boolean
-  }
-}
-
-export interface TipSettingsData {
-  whoami: {
-    tipping: {
-      lastTippedAmount: number
-      limitRemaining: number
-    }
   }
 }
 
@@ -63,20 +43,4 @@ export async function tip(
   }
 
   return tip
-}
-
-export async function tipSettings(
-  this: GraphQlClient,
-  token: string
-): Promise<TipSettingsData['whoami']['tipping']> {
-  const message = await this.query<TipSettingsData>({
-    query: tipSettingsQuery,
-    token
-  })
-
-  if (!message.data?.whoami?.tipping) {
-    throw new Error(`graphql query failed. query=\`${tipSettingsQuery}\``)
-  }
-
-  return message.data?.whoami?.tipping
 }
